Unlicense License?
Hi.

I stumbled across this file, and it seems like a really cool idea to bind chaiscript to ImGui's functions.
I would like to use part of this in a project, however, there's no license information with the code. 

Since it's a really small piece of code, and you put it publicly on the internet, I *assume* you intend to let people *do whatever* with this.

I am proposing to you to add the `unlicense` license to this project. It seem to be the right thing to do. Of course, it's yours, so you do what you want.